Problem: tests fail
Solution: fixed tests,
migrated enclave TX validation to chain-enclave/protocol, such that it can be used
in both mock client and enclave app.
removed the address / witness verification from chain core
and fixed the one in chain-enclave protocol (and moved tests);
added env variable to circle ci for building the enclave
