+++ chain-abci/Cargo.toml
@@ -25,7 +25,6 @@ hex = "0.3"
 protobuf = "2.2.0"
 integer-encoding = "1.0.5"
 clap = { features = ["yaml"], version = "2.32.0" }
-secp256k1zkp = { git = "https://github.com/crypto-com/rust-secp256k1-zkp.git", rev = "99f410ec22c97634e0c430fc578b3a20f30b5ee5" }
 blake2 = "0.8"
 ethbloom = "0.5.0"
 
@@ -34,6 +33,7 @@ quickcheck = "0.8"
 criterion = "0.2"
 digest = "0.8"
 sha3 = "0.8"
+secp256k1zkp = { git = "https://github.com/crypto-com/rust-secp256k1-zkp.git", rev = "99f410ec22c97634e0c430fc578b3a20f30b5ee5" }
 
 [[bench]]
 name = "hashes"
