+++ chain-enclave/enclave-app/src/main.rs
@@ -7,7 +7,9 @@ mod server;
 
 fn main() -> io::Result<()> {
     // TODO: custom runner with args, TLS, remote attestation...
-    let address = env::args().next().unwrap_or("127.0.0.1:7878".to_string());
+    let address = env::args()
+        .next()
+        .unwrap_or_else(|| "127.0.0.1:7878".to_string());
     for stream in TcpListener::bind(address)?.incoming() {
         let mut stream = stream?;
         server::handle_stream(&mut stream);
